scripts/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
Pillow==10.4.0